fixing timestamp conversion

+++ cloudvolume/frontends/graphene.py
@@ -21,6 +21,7 @@ from ..volumecutout import VolumeCutout
 from ..datasource.graphene.metadata import GrapheneApiVersion
 
 from .precomputed import CloudVolumePrecomputed
+import tqdm
 
 def warn(text):
   print(colorize('yellow', text))
@@ -36,7 +37,7 @@ def to_unix_time(timestamp):
   if isinstance(timestamp, datetime): # NB. do not change to elif
     timestamp = datetime.timestamp(timestamp)
 
-  if not isinstance(timestamp, int) and timestamp is not None:
+  if not isinstance(timestamp, float) and timestamp is not None:
     raise ValueError("Not able to convert {} to UNIX time.".format(timestamp))
 
   return timestamp
